chore(web): rename menu buttons to better convey their purpose and implementation
