+++ packages/web/src/views/Forms/EventForm/MigrateBackwardMenuButton.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import styled from "styled-components";
-import { ArrowRight } from "@phosphor-icons/react";
+import { ArrowLeft } from "@phosphor-icons/react";
 import { getMetaKey } from "@web/common/utils/shortcut.util";
 import { Text } from "@web/components/Text";
 import MenuItem from "../ActionsMenu/MenuItem";
 
-const StyledArrowRight = styled(ArrowRight)`
+const StyledArrowLeft = styled(ArrowLeft)`
   width: 16px;
   height: 16px;
 `;
@@ -15,9 +15,9 @@ interface Props {
   tooltipText?: string;
 }
 
-export const MigrateForwardButton: React.FC<Props> = ({
+export const MigrateBackwardMenuButton: React.FC<Props> = ({
   onClick,
-  tooltipText = "Migrate Forward",
+  tooltipText = "Migrate Backward",
 }) => {
   return (
     <MenuItem
@@ -26,11 +26,11 @@ export const MigrateForwardButton: React.FC<Props> = ({
       aria-label={tooltipText}
       tooltipContent={
         <Text size="s" style={{ display: "flex", alignItems: "center" }}>
-          CTRL + {getMetaKey()} + <StyledArrowRight />
+          CTRL + {getMetaKey()} + <StyledArrowLeft />
         </Text>
       }
     >
-      <ArrowRight size={16} />
+      <ArrowLeft size={16} />
       <span>{tooltipText}</span>
     </MenuItem>
   );
