+++ packages/web/src/views/Forms/EventForm/MigrateBelowMenuButton.tsx
@@ -1,37 +1,37 @@
 import React from "react";
 import styled from "styled-components";
-import { ArrowLeft } from "@phosphor-icons/react";
+import { ArrowDown } from "@phosphor-icons/react";
 import { getMetaKey } from "@web/common/utils/shortcut.util";
 import { Text } from "@web/components/Text";
 import MenuItem from "../ActionsMenu/MenuItem";
 
-interface Props {
-  onClick: () => void;
-  label?: string;
-}
-
-const StyledArrowLeft = styled(ArrowLeft)`
-  width: 15px;
-  height: 15px;
+const StyledArrowDown = styled(ArrowDown)`
+  width: 16px;
+  height: 16px;
 `;
 
-export const MoveToSidebarButton: React.FC<Props> = ({
+interface Props {
+  onClick: () => void;
+  tooltipText?: string;
+}
+
+export const MigrateBelowMenuButton: React.FC<Props> = ({
   onClick,
-  label = "Move To Sidebar",
+  tooltipText = "Migrate Below",
 }) => {
   return (
     <MenuItem
       role="menuitem"
       onClick={onClick}
-      aria-label={label}
+      aria-label={tooltipText}
       tooltipContent={
         <Text size="s" style={{ display: "flex", alignItems: "center" }}>
-          CTRL + {getMetaKey()} + <StyledArrowLeft />
+          CTRL + {getMetaKey()} + <StyledArrowDown />
         </Text>
       }
     >
-      <ArrowLeft size={16} />
-      <span>{label}</span>
+      <ArrowDown size={16} />
+      <span>{tooltipText}</span>
     </MenuItem>
   );
 };
