+++ packages/web/src/views/Forms/EventForm/MigrateForwardMenuButton.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import styled from "styled-components";
-import { ArrowUp } from "@phosphor-icons/react";
+import { ArrowRight } from "@phosphor-icons/react";
 import { getMetaKey } from "@web/common/utils/shortcut.util";
 import { Text } from "@web/components/Text";
 import MenuItem from "../ActionsMenu/MenuItem";
 
-const StyledArrowUp = styled(ArrowUp)`
+const StyledArrowRight = styled(ArrowRight)`
   width: 16px;
   height: 16px;
 `;
@@ -15,9 +15,9 @@ interface Props {
   tooltipText?: string;
 }
 
-export const MigrateAboveButton: React.FC<Props> = ({
+export const MigrateForwardMenuButton: React.FC<Props> = ({
   onClick,
-  tooltipText = "Migrate Above",
+  tooltipText = "Migrate Forward",
 }) => {
   return (
     <MenuItem
@@ -26,11 +26,11 @@ export const MigrateAboveButton: React.FC<Props> = ({
       aria-label={tooltipText}
       tooltipContent={
         <Text size="s" style={{ display: "flex", alignItems: "center" }}>
-          CTRL + {getMetaKey()} + <StyledArrowUp />
+          CTRL + {getMetaKey()} + <StyledArrowRight />
         </Text>
       }
     >
-      <ArrowUp size={16} />
+      <ArrowRight size={16} />
       <span>{tooltipText}</span>
     </MenuItem>
   );
