+++ packages/web/src/views/Forms/EventForm/MoveToSidebarMenuButton.tsx
@@ -5,25 +5,25 @@ import { getMetaKey } from "@web/common/utils/shortcut.util";
 import { Text } from "@web/components/Text";
 import MenuItem from "../ActionsMenu/MenuItem";
 
-const StyledArrowLeft = styled(ArrowLeft)`
-  width: 16px;
-  height: 16px;
-`;
-
 interface Props {
   onClick: () => void;
-  tooltipText?: string;
+  label?: string;
 }
 
-export const MigrateBackwardButton: React.FC<Props> = ({
+const StyledArrowLeft = styled(ArrowLeft)`
+  width: 15px;
+  height: 15px;
+`;
+
+export const MoveToSidebarMenuButton: React.FC<Props> = ({
   onClick,
-  tooltipText = "Migrate Backward",
+  label = "Move To Sidebar",
 }) => {
   return (
     <MenuItem
       role="menuitem"
       onClick={onClick}
-      aria-label={tooltipText}
+      aria-label={label}
       tooltipContent={
         <Text size="s" style={{ display: "flex", alignItems: "center" }}>
           CTRL + {getMetaKey()} + <StyledArrowLeft />
@@ -31,7 +31,7 @@ export const MigrateBackwardButton: React.FC<Props> = ({
       }
     >
       <ArrowLeft size={16} />
-      <span>{tooltipText}</span>
+      <span>{label}</span>
     </MenuItem>
   );
 };
